pyproject.toml
[project]
name = "snd-revenue-service"
version = "0.1.0"
description = "The Stand and Deliver Revenue Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "discord.py>=2.7,<3",
  "openai>=1.59,<2",
]

[dependency-groups]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.26,<0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
only-include = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/snd_revenue_service"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = true
